move recording framerate from utilities to mocap helpers
